requirements.txt
django
scrapy
scrapyd
python-scrapyd-api
scrapyd-client
pandas
numpy
pandas
scipy
sklearn
nltk
keras
googletrans
tensorflow





